feat(billing): show plans on the billing page, fix pack and checkout UX
Extract the plan cards out of the pricing page into a shared PlanGrid and
render it on the billing page when the account has no paid subscription. The
billing page previously showed a near-empty "Current plan" card and sent the
customer to /pricing to do the only thing the page is for.

Alongside that, several fixes to the surfaces this touches:

- Only offer checkout for plans that have a price. The API accepts 'pro' and
  rejects everything else, so the Free card's enabled "Get Free" button always
  400'd. Free stays actionable for logged-out visitors, where it is the signup
  CTA and routes to /register without touching Stripe.
- Frame annual as the saving rather than the charge. "Billed $1,440 once a
  year" put the largest number on the card next to the $120 it was explaining.
- Move the annual/monthly toggle's discount badge below the Annual button.
  After "Monthly" it read as though monthly were the discounted option.
- Only show one pack's Buy button as in-flight. isPending is mutation-wide, so
  clicking one pack made all four claim they were opening checkout.
- Restore the authentication guard on the pricing page's dashboard link, which
  was sending public visitors into a protected route.
- Drop the per-request rate and "Best rate" chip from the pack cards.
